Move OutputLogger into sensitivity and make internal
